+++ src/libs/supabase/get-event.ts
@@ -1,13 +1,9 @@
+import { PostgrestError } from '@supabase/supabase-js';
 import { supabase } from '../../services/supabase';
-import { logger } from '../../utils';
+import { Tables } from '../../types/Supabase/DatabaseGenerated';
 
-export const getEvents = async () => {
-    try {
-        const { data, error } = await supabase.from('events').select('*');
+export const getEvents = async (): Promise<{ data: Tables<'Event'>[] | null; error: PostgrestError | null }> => {
+    const { data, error } = await supabase.from('Event').select('*');
 
-        return { data, error };
-    } catch (error) {
-        logger.error('Error fetching event:', error);
-        return { data: null, error };
-    }
+    return { data, error };
 };