+++ src/libs/supabase/get-user.ts
@@ -1,13 +1,11 @@
+import { PostgrestError } from '@supabase/supabase-js';
 import { supabase } from '../../services/supabase';
-import { logger } from '../../utils';
+import { Tables } from '../../types/Supabase/DatabaseGenerated';
 
-export const getUser = async (userId: string) => {
-    try {
-        const { data } = await supabase.from('users').select('*').eq('user_id', userId).single();
+export const getUser = async (
+    userId: string,
+): Promise<{ data: Tables<'User'> | null; error: PostgrestError | null }> => {
+    const { data, error } = await supabase.from('User').select('*').eq('user_id', userId).single();
 
-        return { data };
-    } catch (error) {
-        logger.error('Error fetching user:', error);
-        return { data: null, error };
-    }
+    return { data, error };
 };